+++ clang-tools-extra/test/clang-tidy/Inputs/Headers/dcl51.h
@@ -1,0 +1,6 @@
+#pragma clang system_header
+
+namespace std {
+    #define override __override__
+    int __variable__;
+}